backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
bandit==1.7.5
black==23.7.0
boto3==1.26.143
botocore==1.29.143
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-filter==23.2
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-yaml==2.0.0
djoser==2.2.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==19.1.0
filelock==3.12.0
flake8==6.0.0
gitdb==4.0.10
GitPython==3.1.31
hyperlink==21.0.0
identify==2.5.24
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
jmespath==1.0.1
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
pbr==5.11.1
Pillow==10.0.0
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
pytest==7.4.0
pytest-django==4.5.2
pytest-mock==3.11.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.5
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stevedore==5.1.0
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
urllib3==1.26.16
virtualenv==20.23.0
zope.interface==6.0
